refactor(gui-v2): move expanded form header part to a component
